# scripts/openapi-snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "JobOut": {
    "properties": {
     "attempts": {
      "title": "Attempts",
      "type": "integer"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "error_message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Message"
     },
     "finished_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "heartbeat_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Heartbeat At"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "max_attempts": {
      "title": "Max Attempts",
      "type": "integer"
     },
     "metrics": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Metrics"
     },
     "stage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stage"
     },
     "started_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "status": {
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "id",
     "kind",
     "status",
     "stage",
     "attempts",
     "max_attempts",
     "error_code",
     "error_message",
     "created_at",
     "started_at",
     "finished_at",
     "heartbeat_at"
    ],
    "title": "JobOut",
    "type": "object"
   },
...
   "ObligationCreateRequest": {
    "additionalProperties": false,
    "properties": {
     "claim_row_id": {
      "anyOf": [
       {
        "maxLength": 36,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Claim Row Id"
     },
     "deadline": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline"
     },
     "note": {
      "anyOf": [
       {
        "maxLength": 2000,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "owner": {
      "anyOf": [
       {
        "maxLength": 255,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner"
     }
    },
    "title": "ObligationCreateRequest",
    "type": "object"
   },
   "ObligationOut": {
    "properties": {
     "claim_row_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Claim Row Id"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By"
     },
     "deadline": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "owner": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner"
     },
     "report_id": {
      "title": "Report Id",
      "type": "string"
     },
     "resolved_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Resolved At"
     },
     "status": {
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "id",
     "report_id",
     "claim_row_id",
     "owner",
     "deadline",
     "status",
     "note",
     "created_by",
     "created_at",
     "resolved_at"
    ],
    "title": "ObligationOut",
    "type": "object"
   },
   "ObligationUpdateRequest": {
    "additionalProperties": false,
    "properties": {
     "deadline": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline"
     },
     "note": {
      "anyOf": [
       {
        "maxLength": 2000,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "owner": {
      "anyOf": [
       {
        "maxLength": 255,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner"
     },
     "status": {
      "anyOf": [
       {
        "enum": [
         "OPEN",
         "IN_PROGRESS",
         "RESOLVED"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     }
    },
    "title": "ObligationUpdateRequest",
    "type": "object"
   },
...
   "Page_ObligationOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ObligationOut"
      },
      "title": "Items",
      "type": "array"
     },
     "limit": {
      "title": "Limit",
      "type": "integer"
     },
     "offset": {
      "title": "Offset",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "Page[ObligationOut]",
    "type": "object"
   },
   "Page_ReportOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ReportOut"
      },
      "title": "Items",
      "type": "array"
     },
     "limit": {
      "title": "Limit",
      "type": "integer"
     },
     "offset": {
      "title": "Offset",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "Page[ReportOut]",
    "type": "object"
   },
   "ReportOut": {
    "properties": {
     "coverage_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Coverage Pct"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "expires_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     },
     "file_kind": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "File Kind"
     },
     "file_name": {
      "title": "File Name",
      "type": "string"
     },
     "file_size_bytes": {
      "title": "File Size Bytes",
      "type": "integer"
     },
     "grade": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grade"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "ingest_notes": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ingest Notes"
     },
     "issues_found": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Issues Found"
     },
     "job": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "processing_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Processing Error"
     },
     "programme": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Programme"
     },
     "rows_processed": {
      "title": "Rows Processed",
      "type": "integer"
     },
     "rows_total": {
      "title": "Rows Total",
      "type": "integer"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "sender": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender"
     },
     "sheet_count_total": {
      "title": "Sheet Count Total",
      "type": "integer"
     },
     "source_channel": {
      "default": "upload",
      "title": "Source Channel",
      "type": "string"
     },
     "source_sha256": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Sha256"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "updated_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated At"
     }
    },
    "required": [
     "id",
     "created_at",
     "file_name",
     "file_size_bytes",
     "sheet_count_total",
     "rows_processed",
     "rows_total",
     "coverage_pct",
     "grade",
     "score",
     "status"
    ],
    "title": "ReportOut",
    "type": "object"
   },
...
   "SheetOut": {
    "properties": {
     "fields_mapped": {
      "title": "Fields Mapped",
      "type": "integer"
     },
     "fields_total": {
      "title": "Fields Total",
      "type": "integer"
     },
     "header_row_index": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Header Row Index"
     },
     "hidden": {
      "title": "Hidden",
      "type": "boolean"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "mapping_status": {
      "title": "Mapping Status",
      "type": "string"
     },
     "needs_review": {
      "title": "Needs Review",
      "type": "integer"
     },
     "notes": {
      "items": {
       "type": "string"
      },
      "title": "Notes",
      "type": "array"
     },
     "row_count": {
      "title": "Row Count",
      "type": "integer"
     },
     "sheet_index": {
      "title": "Sheet Index",
      "type": "integer"
     },
     "sheet_name": {
      "title": "Sheet Name",
      "type": "string"
     },
     "skip_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Skip Reason"
     },
     "source_column_count": {
      "title": "Source Column Count",
      "type": "integer"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "trailing_blank_rows": {
      "title": "Trailing Blank Rows",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "sheet_name",
     "sheet_index",
     "header_row_index",
     "row_count",
     "source_column_count",
     "status",
     "skip_reason",
     "hidden",
     "notes",
     "trailing_blank_rows",
     "mapping_status",
     "fields_mapped",
     "fields_total",
     "needs_review"
    ],
    "title": "SheetOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "TrueBind API",
  "version": "0.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/obligations": {
   "get": {
    "operationId": "list_obligations_api_v1_obligations_get",
    "parameters": [
     {
      "in": "query",
      "name": "report_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 36,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Report Id"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "pattern": "^(OPEN|IN_PROGRESS|RESOLVED|OVERDUE)$",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "maximum": 10000000,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ObligationOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Obligations",
    "tags": [
     "obligations"
    ]
   }
  },
  "/api/v1/obligations/{obligation_id}": {
   "patch": {
    "operationId": "update_obligation_api_v1_obligations__obligation_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "obligation_id",
      "required": true,
      "schema": {
       "title": "Obligation Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ObligationUpdateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ObligationOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Obligation",
    "tags": [
     "obligations"
    ]
   }
  },
...
  "/api/v1/reports": {
   "get": {
    "operationId": "list_reports_api_v1_reports_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "maximum": 10000000,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ReportOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Reports",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/obligations": {
   "post": {
    "operationId": "create_obligation_api_v1_reports__report_id__obligations_post",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ObligationCreateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ObligationOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Obligation",
    "tags": [
     "obligations"
    ]
   }
  },
  "/api/v1/reports/{report_id}/process": {
   "post": {
    "operationId": "process_report_api_v1_reports__report_id__process_post",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Process Report",
    "tags": [
     "mapping"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/sheets": {
   "get": {
    "operationId": "list_sheets_api_v1_reports__report_id__sheets_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SheetOut"
         },
         "title": "Response List Sheets Api V1 Reports  Report Id  Sheets Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Sheets",
    "tags": [
     "mapping"
    ]
   }
  },
...
 }
}